Ensure body is given to JsonClient.post
If body is was not defined, body would be null (which passes typeof
test because typeof null == 'object') and JSON.parse(body) would coerce
it to string 'null'. This got sent to the server and triggered
json_body_parser to send back InvalidContentError (Invalid JSON).

JsonClient should throw an error if body is falsy to prevent this.