Fix #472 Use native http/https.Agent
The 'keep-alive-agent' module is not functional on Node v0.11.6 and
above due to http agent class refactoring.  While this breakage is
arguably a bug in Node, there's also no need for a keep alive agent in
Node v0.11 and above, since http keepalive is supported natively now.

Sniff for the Node version, and choose the appropriate agent and configs
accordingly.