next(err) does not call res.send on a finished res
If an upstream handler closes the response successfully (or not), and
a downstream handler passes an unrelated error to `next`, an uncaught
exception will be thrown for trying to set headers that have already been
sent.

This commit alters the behavior of `next` such that its error-handling
path only calls `res.send` if the response is not finished yet.

Cf / closes #1019