fix(static): avoid user-provided data in Error messages being interpreted as sprintf codes (#1384)
The 'static' plugin had a few cases where the path in a request would be
passed as the first ("message") field to a RestError constructor.
RestError uses verror.WError, which uses extsprintf to render the given
arguments. If the "message" includes "%...s" or similar printf codes,
then it will error output.

Also bump to 4.3.1.

* drop ver bump changes, per review comments

* update test name and drop comment (unnecessary with issue ref) per review comments