fix(static): avoid user-provided data in Error messages being interpreted as sprintf codes (#1384) (#1472)
The 'static' plugin had a few cases where the path in a request would be
passed as the first ("message") field to a RestError constructor.
RestError uses verror.WError, which uses extsprintf to render the given
arguments. If the "message" includes "%...s" or similar printf codes,
then it will error output.