Inject `form` to `multipartHandler` functions
Injecting formidable `form` to `multipartHandler` and `multipartFileHandler` functions as the third argument can give us the ability to call `form.handlePart(part)` inside the customized functions. It's useful when the user wants to restrict input files to some extensions e.g. png/jpg/jpeg etc. As a result, when the `part.mime` would be acceptable, the user can continue the process.