fix: make arity error message actionable
The error that is bubbled up when using a handler with the wrong number
of arguments was an unintelligible assertion error with no info about
what failed. This made it nearly impossible to find the problem.

This change adds a better error message including contextual info so
that this error is actionable.