feat: allow custom alternatives to domains
domains is deprecated, but there isn't a standard alternative in the
ecosystem as of right now. Instead of implementing a bespoke alternative
to it, let users create their own alternatives (or use libraries) by
providing a generic way to replace domains in the error handling
workflow. Note: unless you know what you're doing, you should NOT be
using this feature.